Add SchoolSearch class for paying_for_college
